refactor: 사용자인수테스트 메소드 위치 이동
MemberAcceptanceTest -> MemberAcceptanceTestSupport
